refactor: test 폴더와 src 폴더를 루트 디렉토리에 두기 위해 폴더구조 수정 - 기존 코드는 src/하위로 이동
