+++ src/AuthManager.ts
