+++ src/factories/AuthManagerFactory.ts
