+++ src/network/interfaces/HttpClient.ts
