+++ src/network/interfaces/index.ts
