+++ src/providers/interfaces/IEmailVerifiable.ts
