+++ src/providers/interfaces/config/auth-config.ts
