+++ src/shared/types/auth.ts
