+++ src/shared/types/common.ts
