+++ src/shared/types/literals.ts
