+++ src/shared/utils/errorUtils.ts
