+++ src/storage/TokenStore.interface.ts
