+++ tsconfig.json
@@ -5,7 +5,7 @@
     "moduleResolution": "node",
     "lib": ["es2020", "dom"],
     "outDir": "./dist",
-    "rootDir": "./",
+    "rootDir": "./src",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
@@ -17,7 +17,7 @@
     "allowSyntheticDefaultImports": true
   },
   "include": [
-    "**/*.ts"
+    "src/**/*.ts"
   ],
   "exclude": [
     "node_modules",
