+++ vitest.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'node',
-    setupFiles: ['./test/setup.ts'],
+    setupFiles: ['./test/setup.unit.ts'],
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'html'],
